revision_binder правки пра
