+++ mvi-core/src/main/kotlin/ru/surfstudio/mvi/flow/FlowBinder.kt
@@ -42,12 +42,7 @@ interface FlowBinder {
                 Log.e(TAG, it.message, it)
                 throw it
             }.shareIn(this, SharingStarted.Eagerly)
-        this@bind.launch {
-            middleware.transform(eventFlow)
-                .collect { transformedEvent: T ->
-                    eventHub.emit(transformedEvent)
-                }
-        }
+        transformEvents(eventHub, middleware, eventFlow)
     }
 
     fun <T : Event> CoroutineScope.bind(
@@ -59,7 +54,15 @@ interface FlowBinder {
                 Log.e(TAG, it.message, it)
                 throw it
             }.shareIn(this, SharingStarted.Eagerly)
-        this@bind.launch {
+        transformEvents(eventHub, middleware, eventFlow)
+    }
+
+    private fun <T : Event> CoroutineScope.transformEvents(
+        eventHub: FlowEventHub<T>,
+        middleware: Middleware<Flow<T>, Flow<T>>,
+        eventFlow: SharedFlow<T>
+    ) {
+        launch {
             middleware.transform(eventFlow)
                 .collect { transformedEvent: T ->
                     eventHub.emit(transformedEvent)
