+++ mvi-core/src/main/kotlin/ru/surfstudio/mvi/vm/compose/MviBinds.kt
@@ -5,12 +5,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
 import kotlinx.coroutines.launch
 import ru.surfstudio.mvi.core.event.Event
-import ru.surfstudio.mvi.vm.MviViewModelWithoutState
+import ru.surfstudio.mvi.vm.BaseViewModel
 
 /** Syntax sugar fun for convenient binding in @Composable with MVI */
 @SuppressLint("ComposableNaming")
 @Composable
-infix fun <E : Event> MviViewModelWithoutState<E>.binds(
+infix fun <E : Event> BaseViewModel<E>.binds(
     render: @Composable ComposedViewContext<E>.() -> Unit
 ) {
     val scope = rememberCoroutineScope()
