+++ mvi-mappers/src/main/java/ru/surfstudio/mvi/mappers/handler/MviErrorHandlerViewModel.kt
@@ -16,10 +16,10 @@
 package ru.surfstudio.mvi.mappers.handler
 
 import ru.surfstudio.mvi.core.event.Event
-import ru.surfstudio.mvi.vm.MviViewModel
+import ru.surfstudio.mvi.vm.BaseStateFullViewModel
 
-/** More complex [MviViewModel] implementation which reducer is able to handle errors */
-abstract class MviErrorHandlerViewModel<S : Any, E : Event> : MviViewModel<S, E>() {
+/** More complex [BaseStateFullViewModel] implementation which reducer is able to handle errors */
+abstract class MviErrorHandlerViewModel<S : Any, E : Event> : BaseStateFullViewModel<S, E>() {
 
     abstract override val reducer: ErrorHandlerReducer<E, S>
 }
