+++ sample/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
         </activity>
 
         <activity android:name=".handler.HandlerActivity" />
-        <activity android:name=".compose.SimpleComposeActivity" />
+        <activity android:name=".compose.PlaceComposeActivity" />
     </application>
 
 </manifest>
