+++ sample/src/main/kotlin/ru/surfstudio/mvi/flow/app/compose/simple/SimpleComposeEvent.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ru.surfstudio.mvi.flow.app.compose
+package ru.surfstudio.mvi.flow.app.compose.simple
 
 import ru.surfstudio.mvi.core.event.Event
