+++ sample/src/main/kotlin/ru/surfstudio/mvi/flow/app/compose/simple/SimpleViewModel.kt
@@ -1,0 +1,14 @@
+package ru.surfstudio.mvi.flow.app.compose.simple
+
+import ru.surfstudio.mvi.flow.FlowEventHub
+import ru.surfstudio.mvi.vm.BaseViewModel
+
+class SimpleComposeViewModel : BaseViewModel<SimpleComposeEvent>() {
+
+    override val hub: FlowEventHub<SimpleComposeEvent> = FlowEventHub()
+    override val middleware: SimpleComposeMiddleware = SimpleComposeMiddleware()
+
+    init {
+        bindFlow()
+    }
+}
