+++ sample/src/main/kotlin/ru/surfstudio/mvi/flow/app/compose/standard/ComposeScreen.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ru.surfstudio.mvi.flow.app.compose
+package ru.surfstudio.mvi.flow.app.compose.standard
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -26,7 +26,6 @@ import androidx.compose.ui.Modifier
 import androidx.lifecycle.viewmodel.compose.viewModel
 import ru.surfstudio.mvi.flow.app.reused.NetworkEvent
 import ru.surfstudio.mvi.vm.compose.renders
-import ru.surfstudio.mvi.vm.compose.binds
 
 /**
  * Example composable functions with viewModel
@@ -40,23 +39,6 @@ fun ComposeScreen(viewModel: ComposeViewModel = viewModel()) {
                 Text(text = state.loadStateData)
                 Button(onClick = { emit(NetworkEvent.StartLoading) }) {
                     Text("Start compose loading")
-                }
-            }
-        }
-    }
-}
-
-/**
- * Example composable functions with viewModel, but without state
- */
-@Composable
-fun ComposeScreenWithoutState(viewModel: ComposeViewModelWithoutState = viewModel()) {
-    viewModel binds {
-        Surface(modifier = Modifier.fillMaxWidth()) {
-            Column(horizontalAlignment = Alignment.CenterHorizontally) {
-                Text(text = "Compose function with viewModel, but without state")
-                Button(onClick = { emit(SimpleComposeEvent.SimpleClickEventEvent) }) {
-                    Text("Click click! See log")
                 }
             }
         }
